Move trie implementation to new file
